requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bilibili-api-python==16.1.0
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
outcome==1.2.0
Pillow==10.0.1
prettytable==3.9.0
pyasn1==0.5.0
pycryptodomex==3.18.0
pyecharts==2.0.4
pypng==0.20220715.0
PySocks==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
qrcode-terminal==0.8
requests==2.31.0
rsa==4.9
selenium==4.12.0
simplejson==3.19.1
six==1.16.0
snapshot-selenium==0.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
wcwidth==0.2.6
wsproto==1.2.0
yarl==1.9.2
